refactor: split business list pagination code to single components
